+++ src/pyochain/__init__.py
@@ -2,7 +2,7 @@
 
 from . import traits
 from ._dict import Dict
-from ._iter import Iter, Peekable, Unzipped, Vec
+from ._iter import Iter, Vec
 from ._range import Range
 from ._seq import Seq
 from ._set import Set, SetMut
@@ -30,7 +30,6 @@ __all__ = [
     "Ok",
     "Option",
     "OptionUnwrapError",
-    "Peekable",
     "Range",
     "Result",
     "ResultUnwrapError",
@@ -38,7 +37,6 @@ __all__ = [
     "Set",
     "SetMut",
     "Some",
-    "Unzipped",
     "Vec",
     "option",
     "then_if_some",